requirements.txt
gdcm==1.1
imagecodecs==2025.3.30
numpy==2.3.1
pillow==11.3.0
pydicom==3.0.1
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.3.0
pylibjpeg-openjpeg==2.4.0
